data type

+++ file_io/2-append_text_to_file.c
@@ -9,7 +9,7 @@
 int create_file(const char *filename, char *text_content)
 {
 int file_descr;
-size_t bytes;
+int bytes;
 
 if (!filename)
 return (-1);
